Moved TypeBlockWrapper from EditableDisplay/helpers into EditableDisplay/types/helpers.
