+++ app/modules/contentItems/components/EditableDisplay/types/Heading.js
@@ -7,7 +7,8 @@ import type { Identifier } from 'types/model';
 import EditableTextContent from 'core-components/EditableTextContent';
 
 import * as m from '../../../model';
-import TypeBlockWrapper from '../helpers/TypeBlockWrapper';
+
+import TypeBlockWrapper from './helpers/TypeBlockWrapper';
 
 import { passThroughProps } from '..';
